kubevirt: split VMI selectors
Hot fix to enable integration tests.

This change does not break other flows, just enables to import simplier
modules to by-pass immediate execution of code in more complex ones.
