+++ frontend/packages/kubevirt-plugin/src/selectors/vmi/index.ts
@@ -1,2 +1,3 @@
+export * from './basic';
 export * from './combined';
 export * from './selectors';
